pyproject.toml
[tool.poetry]
name = "roundcube-session-decrypt"
version = "0.1.0"
description = ""
authors = ["Raphael Dray <dray.raph@gmail.com>"]
readme = "README.md"
packages = [{include = "roundcube_session_decrypt"}]

[tool.poetry.dependencies]
python = "^3.11"
phpserialize = "^1.3"
pycryptodome = "^3.23.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
